+++ Flow.Plugin.WindowWalker.csproj
@@ -68,12 +68,12 @@
 
     <ItemGroup>
         <PackageReference Include="Flow.Launcher.Plugin" Version="2.1.0" />
-        <PackageReference Include="JetBrains.Annotations" Version="2021.3.0" />
+        <PackageReference Include="JetBrains.Annotations" Version="2022.1.0" />
         <PackageReference Include="Microsoft.CodeAnalysis.NetAnalyzers" Version="6.0.0">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Microsoft.Windows.CsWin32" Version="0.1.691-beta">
+        <PackageReference Include="Microsoft.Windows.CsWin32" Version="0.2.1-beta">
             <PrivateAssets>all</PrivateAssets>
         </PackageReference>
         <PackageReference Include="Microsoft.Windows.SDK.Win32Metadata" Version="23.0.2-preview" />
